IMPALA-13993: waitForHmsEvent should check table events under missing dbs
When a db is missing in the catalog cache, waitForHmsEvent request
currently just check if there are pending database events on it,
assuming processing the CREATE_DATABASE event will add the db with
the table list. However, that's wrong since processing CREATE_DATABASE
event just adds the db with an empty table list. We should still wait
for pending events on underlying tables.

Tests:
 - Added e2e test which consistenly fails when running concurrently with
   other tests in TestEventSyncWaiting without the fix.

Change-Id: I3fe74fcf0bf4dbac4a3584f6603279c0a2730b0c
Reviewed-on: http://gerrit.cloudera.org:8080/22817
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Quanlong Huang <huangquanlong@gmail.com>