IMPALA-13999: Refactor test_hms_event_sync_basic to be smaller parallel tests
test_hms_event_sync_basic is not a simple test that it actually tests
several kinds of statements in sequence.

This refactors it into smaller parallel tests so there are more
concurrent HMS events to be processed and easier to reveal bugs.

Renamed some tests to use shorter names.

Tests:
 - Ran all parallel tests of TestEventSyncWaiting 32 times.

Change-Id: I8a2be548697f6259961b83dc91230306f38e03ad
Reviewed-on: http://gerrit.cloudera.org:8080/22829
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>