IMPALA-14029: Add Kerberos utilities to Docker image build
The Kerberos utility package was missing from the OS package list of
the Docker container build when the base image was detected being
a hardened Wolfi-based image. This prevented Impala coordinators from
renewing their Kerberos tickets in containerized and Kerberized
environments.

This patch adds the Kerberos utility package to the list of installed
packages for such minimal containers.

Change-Id: I84f295ac8ae4c000868abff0342b922beb141b5b
Reviewed-on: http://gerrit.cloudera.org:8080/22854
Reviewed-by: Norbert Luksa <norbert.luksa@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>