IMPALA-14072: Fix NPE in Catalogd during rename.
test_rename_drop fail with NPE after IMPALA-14042. This is because
CatalogServiceCatalog.renameTable() return null for not finding the
database of oldTableName. This patch change renameTable() to return
Pair.create(null, null) for that scenario.

Refactor test_rename_drop slightly to ensure that invalidating the
renamed table and dropping it are successful.

Testing:
- Add checkNotNull precondition in
  CatalogOpExecutor.alterTableOrViewRename()
- Increase catalogd_table_rename_delay delay to 6s to ensure that DROP
  query happen in Catalogd before renameTable() called. Manually
  observed that No NPE is shown anymore.

Change-Id: I7a421a71cf3703290645e85180de8e9d5e86368a
Reviewed-on: http://gerrit.cloudera.org:8080/22899
Reviewed-by: Quanlong Huang <huangquanlong@gmail.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>