IMPALA-14066 (Part 6): Re-applying IMPALA-14038: Pull in KUDU-3663 to handle certs with RSASSA-PSS
This commit re-applies IMPALA-14038 to the Kudu files after the Kudu
rebase to v1.17.1.

Testing:
 - exhaustive tests have passed

The original commit message is below:

The existing KRPC code to determine the hash algorithm for a
certificate does not handle RSASSA-PSS signatures as the hash
algorithm is configurable for RSASSA-PSS. This was addressed
in Kudu with KUDU-3663. That fix uses OpenSSL 1.1.1's
x509_get_signature_info() function, which is able to determine
the hash algorithm even for RSASSA-PSS. This is similar to the
fix that Postgres did in a similar situation. It does not support
RSASSA-PSS on OpenSSL 1.0.2, but it improves the error message
in that case.

Testing:
 - Kudu added a unit test that passes

Change-Id: I162efac2d68c2bfb34a5086557182f68670d8c2a
Reviewed-on: http://gerrit.cloudera.org:8080/22923
Reviewed-by: Jason Fehr <jfehr@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-on: http://gerrit.cloudera.org:8080/23015
Reviewed-by: Daniel Becker <daniel.becker@cloudera.com>
Tested-by: Daniel Becker <daniel.becker@cloudera.com>