IMPALA-14042: (Addendum) limit test_rename_drop in exhaustive mode
test_rename_drop continues to be flaky in non-HDFS environment. HMS
seems to be slower to response in non-HDFS environment. There are also
possibillity of table lock contention with background TableLoader and
CatalogDelta thread that cause DROP query happen after RENAME instead of
BEFORE.

This patch limit running test_rename_drop in HDFS and exhaustive mode in
the meantime.

Change-Id: Ie55e6d6093367c454cf3e31ed8a409b6e091193d
Reviewed-on: http://gerrit.cloudera.org:8080/22933
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Michael Smith <michael.smith@cloudera.com>