IMPALA-14102: [part 1] Calcite Planner: optimize join rule
This is part 1 of the commit for optimizing join rules for
Calcite.  This commit is just a copy of the LoptOptimizeJoinRule.java
from Calcite v1.37 for subsequent modification.

The purpose of this commit is to serve as a placeholder
starting point so we can easily see the customized changes that
are made by comparing Impala specific modifications for the rule
which will be done in subsequent commits for IMPALA-14102.

Change-Id: I63daf6dacf0547a0488c1ecf0bc185b548e00d87
Reviewed-on: http://gerrit.cloudera.org:8080/23312
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Aman Sinha <amsinha@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>