IMPALA-14109: Remove SkipIfCatalogV2.hms_event_polling_disabled
This skipIf used the coordinator webui to check whether the flag
is set and skipped the test if the cluster was not running.
The skipIf was only used in custom cluster tests where the cluster
is restarted with new flags anyway, so the flags of the previous
cluster are not relevant.

Change-Id: I455b39eff95e45d02c7b9e0b35d8e7fe03145bb1
Reviewed-on: http://gerrit.cloudera.org:8080/22960
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>