IMPALA-14041 ADDENDUM: Recreate jar file after tests run.
The build was unstable because the "mvn -fae" command in
the java/calcite-planner was removing the jar file due to
the "clean" plugin which was needed because of a mvn bug.

This caused the custom_cluster test_calcite_planner.py test to
fail because it wouldn't find the jar file.

Adding the "install" ensures the jar file will be recreated
after the tests are run.

Change-Id: I04e2c1b9bef26535ca571b87fdbd3e84c61b4b24
Reviewed-on: http://gerrit.cloudera.org:8080/22982
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>